fix(ui): navbar style, hydration and ssr issues (issue 4)
