+++ src/app/restaurants/[restaurantName]/error.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import Button from '@/components/ui/Button';
 import Link from 'next/link';
 import { useEffect } from 'react';
+import { cn } from '@/lib/utils';
+import { buttonVariants } from '@/components/ui/Button';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const Error = ({ error, reset }: { error: Error; reset: () => void }) => {
@@ -15,9 +16,12 @@ const Error = ({ error, reset }: { error: Error; reset: () => void }) => {
       <h1 className="mb-5 text-5xl">Oops!</h1>
       <p>Something went wrong!</p>
       <p>{error.message}</p>
-      <Button variant="link" className="p-0">
-        <Link href="/restaurants">Return to Restaurants page</Link>
-      </Button>
+      <Link
+        href="/restaurants"
+        className={cn('p-0', buttonVariants({ variant: 'link' }))}
+      >
+        Return to Restaurants page
+      </Link>
     </>
   );
 };
